+++ src/model/mod.rs
@@ -1,5 +1,7 @@
 mod error;
+mod session;
 mod shell_file;
 
 pub use error::*;
+pub use session::*;
 pub use shell_file::*;